feat: implement logging middleware with rate limiting for login attempts and payment redirection flow

+++ back/app/modules/producto/schemas.py
@@ -1,17 +1,18 @@
 from typing import Optional, List
 from sqlmodel import SQLModel, Field
+from decimal import Decimal
 from app.modules.categoria.schemas import CategoriaPublic
 from app.modules.ingerediente.schemas import IngredientePublic
 from app.modules.unidadMedida.schemas import UnidadMedidaPublic
 
 
-
 # ── Entrada ───────────────────────────────────────────────────────────────────
+
 
 class ProductoCreate(SQLModel):
     nombre: str = Field(min_length=2, max_length=100)
     descripcion: str = Field(min_length=2, max_length=500)
-    precio_base: int = Field(ge=0)
+    precio_base: Decimal = Field(ge=0)
     imagen_url: str = Field(max_length=255)
     unidad_venta_id: Optional[int] = Field(default=None)
 
@@ -19,7 +20,7 @@ class ProductoCreate(SQLModel):
 class ProductoUpdate(SQLModel):
     nombre: Optional[str] = Field(default=None, min_length=2, max_length=100)
     descripcion: Optional[str] = Field(default=None, min_length=2, max_length=500)
-    precio_base: Optional[int] = Field(default=None, ge=0)
+    precio_base: Optional[Decimal] = Field(default=None, ge=0)
     imagen_url: Optional[str] = Field(default=None, max_length=255)
     is_active: Optional[bool] = None
     unidad_venta_id: Optional[int] = Field(default=None)
@@ -28,8 +29,10 @@ class ProductoUpdate(SQLModel):
 class CategoriaAssign(SQLModel):
     categoria_ids: List[int] = Field(default_factory=list)
 
+
 class IngredienteAssign(SQLModel):
     ingrediente_ids: List[int] = Field(default_factory=list)
+
 
 class UnidadMedidaAssign(SQLModel):
     unidad_venta_id: Optional[int] = Field(default=None)
@@ -41,7 +44,7 @@ class ProductoPublic(SQLModel):
     id: int
     nombre: str
     descripcion: str
-    precio_base: int
+    precio_base: Decimal
     imagen_url: str
     is_active: bool
     unidad_venta_id: Optional[int] = None
@@ -54,4 +57,3 @@ class ProductoPublic(SQLModel):
 class ProductoList(SQLModel):
     data: List[ProductoPublic]
     total: int
-
